requirements.txt
paho-mqtt
bleak==0.20.5
